requirements.txt
fastapi
pillow
uvicorn
requests
numpy
pydantic